+++ Changelog.md
@@ -2,6 +2,7 @@ AirportBrcmFixup Changelog
 ============================
 #### v2.1.5
 - Fix compilation issues in Xcode 13.3.1
+- Fix patching of AirPortBrcmNIC for Monterey (kext located in IO80211FamilyLegacy.kext/PlugIns)
 
 #### v2.1.4
 - boot-arg and property `brcmfx-delay` (Number) delays start of native broadcom driver for specified amount of milliseconds. Can be required for successful start in macOS Monterey 12.1 and newer versions.